+++ src/utils/auth_utils.ts
@@ -3,19 +3,14 @@
 import fs from "fs";
 import vscode from 'vscode';
 import yaml from 'js-yaml';
-import express from "express";
 import detectPort from "detect-port";
 
-import { readYML } from './common';
-import {
-    Auth0URLs,
-    LOGIN_SUCCESS_CALLBACK,
-    NOTIFICATION
-} from "../constants";
-import { repoIsNotSynced } from "../events/utils";
-import { showConnectRepo } from "./notifications";
-import { createUserWithApi } from "./api_utils";
-import { generateSettings } from "../settings";
+import {readYML} from './common';
+import { Auth0URLs, NOTIFICATION } from "../constants";
+import {repoIsNotSynced} from "../events/utils";
+import {showConnectRepo} from "./notifications";
+import {createUserWithApi} from "./api_utils";
+import {generateSettings} from "../settings";
 
 
 export const isPortAvailable = async (port: number) => {
@@ -29,32 +24,9 @@ export const isPortAvailable = async (port: number) => {
         });
 };
 
-export const initExpressServer = () => {
-    // Create an express server
-    const expressApp = express();
-    const port = (global as any).port;
-
-    // define a route handler for the default home page
-    expressApp.get("/", async (req: any, res: any) => {
-        res.send("OK");
-    });
-
-    // define a route handler for the authorization callback
-    expressApp.get(LOGIN_SUCCESS_CALLBACK, async (req: any, res: any) => {
-        const repoPath = vscode.workspace.rootPath || "";
-        await createUser(req.query.access_token, req.query.id_token, repoPath);
-        res.send(NOTIFICATION.LOGIN_SUCCESS);
-    });
-
-    // start the Express server
-    expressApp.listen(port, () => {
-        console.log(`server started at ${port}`);
-    });
-};
-
 export const createRedirectUri = () => {
     const port = (global as any).port;
-    return `http://localhost:${port}${LOGIN_SUCCESS_CALLBACK}`;
+    return `http://localhost:${port}${Auth0URLs.LOGIN_CALLBACK_PATH}`;
 };
 
 export const redirectToBrowser = (skipAskConnect=false) => {