+++ src/utils/setup_utils.ts
@@ -7,12 +7,13 @@ import {
 	NOTIFICATION
 } from "../constants";
 import { repoIsNotSynced } from '../events/utils';
-import { initExpressServer, isPortAvailable } from './auth_utils';
+import { isPortAvailable } from './auth_utils';
 import { showConnectRepo, showSignUpButtons } from './notifications';
 import { readYML } from './common';
 import { initUtils } from '../init/utils';
 import { trackRepoHandler, unSyncHandler } from '../handlers/commands_handler';
 import { generateSettings } from "../settings";
+import { initExpressServer } from "../server/server";
 
 
 export const createSystemDirectories = () => {